# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.0.0",
    "title": "DREAM Lab API",
    "description": "Endpoints para la API de DREAM Lab. Puede encontrar más información en [GitHub](https://github.com/CodeCraft-Solutions-DREAM-Lab).",
    "contact": {
      "name": "CodeCraft Solutions",
      "url": "https://github.com/CodeCraft-Solutions-DREAM-Lab"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    },
    {
      "url": "https://dreamlab-api.azurewebsites.net/",
      "description": "Live server"
    }
  ],
  "paths": {
...
    "/salas/": {
      "get": {
        "tags": [
          "Salas"
        ],
        "summary": "Obtiene todas las salas",
        "description": "Obtiene todas las salas",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idSala": {
                        "type": "integer"
                      },
                      "nombre": {
                        "type": "string"
                      },
                      "cantidadMesas": {
                        "type": "integer"
                      },
                      "descripcion": {
                        "type": "string"
                      },
                      "fotoURL": {
                        "type": "string"
                      },
                      "detallesURL": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/salas/salasActivas": {
      "get": {
        "tags": [
          "Salas"
        ],
        "summary": "Obtiene las salas que no se encuentren bloqueadas",
        "description": "Obtiene las salas que no se encuentren bloqueadas",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idSala": {
                        "type": "integer"
                      },
                      "nombre": {
                        "type": "string"
                      },
                      "cantidadMesas": {
                        "type": "integer"
                      },
                      "descripcion": {
                        "type": "string"
                      },
                      "fotoURL": {
                        "type": "string"
                      },
                      "detallesURL": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/salas/cronograma": {
      "get": {
        "tags": [
          "Salas"
        ],
        "summary": "Obtiene el cronograma de salas y mesas intercaladas",
        "description": "Obtiene el cronograma de salas y mesas intercaladas",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "Identificador único de la sala o mesa"
                      },
                      "title": {
                        "type": "string",
                        "description": "Nombre de la sala o mesa"
                      },
                      "sala": {
                        "type": "boolean",
                        "description": "Indica si es una sala (true) o una mesa (false)"
                      },
                      "idSala": {
                        "type": "integer",
                        "description": "Identificador de la sala a la que pertenece la mesa"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/salas/horasLibres": {
      "post": {
        "tags": [
          "Salas"
        ],
        "summary": "Obtiene las horas libres de una sala",
        "description": "Obtiene las horas libres de una sala",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "hora": {
                        "type": "integer"
                      },
                      "cupos": {
                        "type": "integer"
                      },
                      "competidores": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idSala": {
                    "type": "integer"
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date"
                  },
                  "personas": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/salas/cambiarEstadoSalas": {
      "put": {
        "tags": [
          "Salas"
        ],
        "summary": "Cambia el estado de bloqueada de una sala",
        "description": "Cambia el estado de bloqueada de una sala de true a false y viceversa",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rowsAffected": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idSala": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/salas/{id}": {
      "get": {
        "tags": [
          "Salas"
        ],
        "summary": "Obtiene una sala por su ID",
        "description": "Obtiene una sala por su ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la sala"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idSala": {
                        "type": "integer"
                      },
                      "nombre": {
                        "type": "string"
                      },
                      "cantidadMesas": {
                        "type": "integer"
                      },
                      "descripcion": {
                        "type": "string"
                      },
                      "fotoURL": {
                        "type": "string"
                      },
                      "detallesURL": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/salas/nameFromExperienceId/{id}": {
      "get": {
        "tags": [
          "Salas"
        ],
        "summary": "Regresa el nombre de una sala basado en el ID de la experiencia",
        "description": "Regresa el nombre de una sala basado en el ID de la experiencia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la experiencia",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "Electric Garage"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}